pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-course-rag-benchmark"
version = "0.1.0"
description = "Open multilingual benchmark for retrieval-grounded educational assistants"
requires-python = ">=3.10"
license = {text = "CC BY-NC-SA 4.0"}
dependencies = [
  "beautifulsoup4>=4.12.0",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.0.0",
  "PyYAML>=6.0",
  "rank-bm25>=0.2.2",
  "requests>=2.31.0",
  "scikit-learn>=1.3.0",
  "scipy>=1.11.0",
  "seaborn>=0.13.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
models = [
  "faiss-cpu>=1.7.4",
  "sentence-transformers>=3.0.0",
  "torch>=2.1.0",
  "transformers>=4.44.0",
]
dev = [
  "pytest>=8.2",
]

[project.scripts]
ocrb = "open_course_rag_benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
